New set for offical release tarballs
* patches/OOO_1_1_RC3/apply: new set for offical release tarballs

* Add patches from Debian packaging:

* patches/RC3/config-switches: add --with-python, --disable-rpath
and move in --without-fonts options

* patches/RC3/config-aclocal.diff: update aclocal, for AM_PATH_PYTHON

* patches/RC3/debug-ooo.diff: set OOO_DEBUG variable to start
soffice.bin with e.g. gdb or strace

* patches/RC3/disable-lzwc.diff: hack to bin patentented LZW
compression algorithm for .GIFs.  Needed for Debian

* patches/RC3/dmake-tempnam.diff: fix dmake tempfile vulnerability

* patches/RC3/icu-no-rpath.diff: ICU was patched to include rpaths,
bin it to comply with Debian policy

* patches/RC3/setup-no-langcheck.diff: On upgrade, setup checks that
the install language matches, bin it because we're using langpacks

* patches/RC3/spell-sv-HTML.diff: Fix small typo in Swedish
translation

* patches/RC3/stlport-dynamic-link.diff: Don't link setup statically
to stlport library since we run setup during packaging when
libstlport is available

* patches/RC3/system-python.diff: Initial work to use the system
python instead of the OOo internal version.  This needs more work:
the internal build drops the version suffix and that causes problems

* patches/RC3/system-vera-fonts.diff: (updated) move config switch to
separate patch, don't even unpack the fonts into the solver, don't
pack them for setup.

* patches/apply.pl: Die if PATCHPATH is not set in apply