Patches/RC3_030729/apply:
patches/RC3_030729/apply:
* patches/OOO_1_1_RC3/apply: update comments.

* TODO: upd.

* patches/RC3/zoom-combobox.diff: Add the zoom combobox to the
default functionbar (top toolbar).  Had to copy the definition of
SID_ATTR_ZOOM to sfx2.  What a hack.  And I'm not the first one to
do this, grep for SID_SVX_START in sfxsids.hrc :-(.  Remove the
other zoom combobox from the page preview toolbar in writer.