If java is disabled with --disable-java, don't look for a JDK, but do look for xsltproc, and set XSLTPROC in set_soenv.
* Support building without a JDK.

* allow-no-jdk.diff:  If java is disabled with --disable-java,
don't look for a JDK, but do look for xsltproc, and set
XSLTPROC in set_soenv.

* solenv-set-solar-java-only-if-unset.diff:  Set the SOLAR_JAVA
variable conditionally with "*=".

* readlicense-oo-xsltproc.diff:  Use xsltproc to replace java
for xml transforms in readlicense_oo.

* officecfg-xsltproc.diff:  Use xsltproc to replace java for
xml transforms in officecfg.

* berkeleydb-handle-no-solar-java.diff:  If SOLAR_JAVA is
not set, don't pass --enable-java to the configure script,
and don't expect the jar when finished.

* ridljar-handle-no-solar-java.diff:  If SOLAR_JAVA is not
set, don't attempt to zip up the compiled java files,
since they will not be built.

* jurt-handle-no-solar-java.diff:  Don't build jpipe with
SOLAR_JAVA not set.

* jvmaccess-handle-no-solar-java.diff:	Don't build jvmaccess
with SOLAR_JAVA not set.

* jni-uno-handle-no-solar-java.diff:  Don't build jni_uno
with SOLAR_JAVA not set.

* javaunohelper-handle-no-solar-java.diff:  Don't build
javaunohelper with SOLAR_JAVA not set

* psprint-handle-no-solar-java.diff:  Don't use jvmaccess to
get JDK font path with SOLAR_JAVA not set.

* sj2-handle-no-solar-java.diff:  When SOLAR_JAVA is not
set, replace java applet support with stubs that throw
an exception.

* setup2-handle-no-solar-java.diff:  Don't build jvmsetup or
use jvmaccess for java-related setup with SOLAR_JAVA not set.

* scp-handle-no-solar-java.diff:  Don't ship various
java-related files that will no longer be built, but ship
the stubbed out libj from sj2 because setup fails without it.

* svg-handle-no-solar-java.diff:  Disable SVG import support
with SOLAR_JAVA not set, since it depends on compile-time
and run-time java support.

* xsltvalidate-handle-no-solar-java.diff:  Disable java
components with SOLAR_JAVA not set.

* xsltfilter-handle-no-solar-java.diff:  Disable java
components with SOLAR_JAVA not set.

* desktop-handle-no-solar-java.diff:  Don't build javaldx
with SOLAR_JAVA not set, and check for javaldx in soffice.sh
before using it.