Add --with-jdk-home option that allows you to specify a JDK on your current system rather than use the downloaded one.
* configure.in: add --with-jdk-home option that
allows you to specify a JDK on your current system
rather than use the downloaded one.  Sets the
JDK_HOME variable on ./configure

* bin/build-ooo: various changes to see whether the
user specified --with-jdk-home and use that as
opposed to the downloaded JDK tarball.

* bin/setup.in: add JDK_HOME to PATH if it exists,
and add --with-jdk-home=$JDK_HOME to the OOo
configure script options through CONFIGURE_OPTIONS
if user specified --with-jdk-home at build-ooo
configure time