2004-02-13  Dan Williams <dcbw@redhat.com>
* patches/OOO_1_1/gtk-themeing.diff
patches/OOO_1_1_0/gtk-themeing.diff

Instead of creating a GtkMenuItem widget, also
create a GtkMenu, embed the menu item inside the
menu, then pull the style background color from the
menu item and the text prelight color from the
menu item's label, rather than both from the item.
This fixes behavior of themes like Mist which expect
the label to be in a menu item in a menu.