New configure switch --with-num-jobs.
* configure.in,
bin/build-ooo,
bin/setup.in: New configure switch --with-num-jobs. When specified,
it is used as a -P_number_ parameter for dmake. This allows
distributed building using tools like icecream or distcc. Usual
scenario:

$ export CC='/opt/icecream/bin/gcc'
$ export CXX='/opt/icecream/bin/g++'
$ ./configure --with-ccache-allowed --with-num-jobs=10

Note: --with-ccache-allowed is needed to preserve the CC and CXX
setting.

* patches/OOO_1_1/build-pl-parallel-dmake-blacklist.diff,
patches/OOO_1_1_2/apply,
patches/OOO_1_1_3/apply: A workaround for directories/projects that
are unable to build with "dmake -P_number_"; the -P is removed for
them.

* patches/OOO_1_1/build-pl-conditionals.diff: Fixed typo.