Handle the startup notification in KDE/Gtk+ plugins their native way so that it builds & works fine even without libstartup-notification library.
* patches/src680/vcl-startup-notification.diff,
patches/src680/apply: Handle the startup notification in KDE/Gtk+
plugins their native way so that it builds & works fine even without
libstartup-notification library. It is still recommended to build
with --enable-libsn though; otherwise the startup notification is
confused with the generic plugin when having StartupNotify=true in
the .desktop file.