Small script to update pythonloader.
* bin/pyunorc-update64.in, bin/Makefile.am, bin/package-ooo: small
script to update pythonloader.unorc, so the 32-bit pyuno is
able to work on x86_64. We call the script by rpm trigger when
the python package is installed. I am not sure if it is universal.
I might install it just for Novell but it might be useful for
others as well, so...

* patches/src680/pymailmerge.diff: fix to work with UTF-8 locales,
it seems to work even without non-UTF-8 locales, thanks to
Caolan McNamara