Fix two minor problems in the evoab2 connectivity:  A WHERE expressions of the form field	like 'foo%' was translated into the Evolution addressbook sexp (endswith field "foo") while it should be translated into (beginswith field "foo") and correspondingly for '%foo'.
* patches/evo2/evo2-like-except.diff: Fix two minor problems in
the evoab2 connectivity:

A WHERE expressions of the form
field	like 'foo%'
was translated into the Evolution addressbook sexp
(endswith field "foo")
while it should be translated into
(beginswith field "foo")
and correspondingly for '%foo'.

The handling of LIKE patterns that can't be handled by the
evoab2 driver (a wildcard in the middle of the pattern, or
several wildcards) was wrong.