Check that textures are not empty (doOperation): fix textures handling, before I misunderstood the documentation and thought there's one texture per polygon.
* build/src680-m152/canvas/source/cairo/cairo_canvashelper.cxx
(drawPolyPolygonImplementation): check that textures are not empty
(doOperation): fix textures handling, before I misunderstood the
documentation and thought there's one texture per polygon. In real
it is sequence of textures which is used for all polygons. So far
we use only 1st texture from sequence - same as vcl canvas does
for now.