On Win32, pass --with-build-version with a string indicating the build date.
* bin/setup.in: On Win32, pass --with-build-version with a string
indicating the build date. (Well, the date bin/setup was run,
which might not be the same date as OOo actually was built, if one
then uses the "build" command. Also note that building OOo might
tave several days. But at least it makes it easier to identify a
build based on the Help:About dialog.)