Added new files for external UNO algorithm(s).
* source/inc/numeric/lpuno.hxx:
* source/numeric/lpuno.cxx: added new files for external UNO algorithm(s).

* source/inc/numeric/lpmodel.hxx:
* source/numeric/lpmodel.cxx:
* source/numeric/lpsimplex.cxx: solve(), getSolution() and setSolution removed
from the model class.  So, instead of the model class taking an algorithm object
to find a solution, we now set the model object to the algorithm object, then
have the algorithm object find a solution.  It's simpler this way, and is
consistent with the way the UNO models work.

* source/ui/solvemodel.cxx: altered code due to the above change.