Add a PlainWin32 "distro" that doesn't include the NovellOnly, NovellOnlyWin32, Lockdown or CalcSolver that are in NovellWin32.
* patches/src680/apply: Add a PlainWin32 "distro" that doesn't
include the NovellOnly, NovellOnlyWin32, Lockdown or CalcSolver
that are in NovellWin32. Put system-xmlsec.diff in BuildBits
instead of SystemBits (which is Linux-only) so it gets applied
also on Win32. Not that we ever use any "system" xmlsec on Win32,
but to avoid pain as other diffs depend on this one having been
applied.