Use a Novell-specific
* patches/src680/novell-win32-registry-branding.diff: Use a
Novell-specific *key* pathname (with "Novell" as a component) as
that is what the Novell ZAM people want. Not just a *value*
containing the vendor. Unfortunately this means this patch now
hardcodes "Novell" but upstream doesn't seem to be particularily
interested in this class of functionality anyway.

It would be possible to use the OOO_VENDOR string as subkey
pathname component by adding handling of appropriate metasyntax to
installer::windows::registry::get_registry_key, but that is a bit
messy as that string typically contains punctuation (for us,
"Novell, Inc."), and using such in Registry key pathnames is
highly unusual. So we would need another configure option,
--with-vendor-clean or whatever, and pass just "Novell" for that.