Don't use cairo_helper.
* patches/cairo/cairo-canvas-win32.diff: Don't use
cairo_helper.cxx on Win32 (ifdef out all of it). It's much cleaner
to just have the necessary code inline in cairo_cairo.cxx, which
is the only source file that used the cairo_helper functions
anyway. Additional cleanups. Plug HDC leak: keep HDC and its
original bitmap around in ::cairo::Surface so that we can properly
deallocate the HDC in the destructor.

* patches/cairo/cairocanvas-check-bitmap-system-data-pixmap-size.diff:
Revert my changes on 2006-09-27, not needed now when not using
cairo_helper on Win32..