More work on ODMA support: A new file picker, that is used if DMSes are present on the machine.
* patches/src680/novell-win32-odma.diff: More work on ODMA
support: A new file picker, that is used if DMSes are present on
the machine. If the uses dismisses the DMS's document
creation/selection dialog and wants to "use the application
dialog", the ODMA file picker acts as a proxy for the system file
picker.

Also some essential fixes to the ODMA UCP.

Still need to add an option to disable bothering with ODMA even if
there are DMSes on the machine.