More hacking.
* patches/src680/novell-win32-odma.diff: More hacking. Use the
implementation name when creating the proxied instance of the real
Win32 file picker. Otherwise we could get either another instance
of the ODMA file picker (etc ad nauseam, i.e. stack overflow), or
the system file picker which we want, just depending on what the
order of implementations of the service happened to be in the
services database. The problem is caused by the ODMA file picker
also having to claim it provides the SystemFilePicker service, see
filedlghelper.cxx.

Added a checkbutton on the Tools:Options:General dialog page to
disable showing the DMS dialog. For some reason the button is
insensitive, though. Keep this option in the SvtMiscOptions class.

* patches/src680/apply: Move novell-win32-odma.diff last because
of patch interdependency.