It is too tight to have two checkboxes on the same row (there might be quite verbose localisations of the labels), says erAck.
* patches/src680/novell-win32-odma.diff: It is too tight to have
two checkboxes on the same row (there might be quite verbose
localisations of the labels), says erAck. So put the ODMA dialogs
checkbox on a row of its own. On Unix, where the ODMA dialogs
checkbox isn't displayed, move the controls below it up one
row. If OOo dialogs are used, disable the ODMA dialogs checkbox.