Move fpicker-common-build.
* patches/src680/apply: Move fpicker-common-build.diff from
LinuxOnly to Fpickers, as fpicker-kde-resmgr.diff depends on
it. This means Fpickers can be applied on Win32 too. (Not that
Fpickers contains any code that would be used on Win32, but so
that the KDE file picker sources get expanded and it's easier to
have a look at them. And anyway, it's always better to have
patches tested cross-platform, as that is how they will be used
when eventually upstreamed.)