Use the same VRB fileversion mangling style as in the 2.
* patches/src680/novell-win32-vrb-branding.diff
(desktop/util/ooverinfo.rc solenv/inc/shlinfo.rc): Use the same
VRB fileversion mangling style as in the 2.0.4 branch. Don't
bother with any PATCHLEVEL here, though, so it is quite silly. But
we don't want the fileversions to effectively decrease either, so
we have to do it. See the long comment this diff adds to
ooverinfo.rc.