Use a VRB fileversion that results in increased version compared to that used in the 2.
* patches/src680/novell-win32-vrb-branding.diff
(desktop/util/ooverinfo.rc solenv/inc/shlinfo.rc): Use a VRB
fileversion that results in increased version compared to that
used in the 2.0.4 branch. Don't bother with any PATCHLEVEL here,
though. Can't use the exact same mechanism we *thouht* we were
using in 2.0.4, though... It turns out that the Microsoft resource
compiler is very braindamaged, and silently just ignores attempts
to use multiplication in expressions for the parts in the
fileversion statement. So the minor*100+micro that we used in
2.0.4 didn't actually work, it just seemed to work, because minor
was zero... When minor is 1, minor*100+micro produces the same as
minor+micro. My head spins.

So forget that arithmetic, and just use the VERVARIANT
(i.e. upstream "build id", i.e. the number that was 9073 in 2.0.4
and is 9095 in 2.1.0) as the first part, followed by major, minor,
and micro. See the long comment this diff adds to ooverinfo.rc.