//bugs.
* ported the following three fixes from the ooo-build-2-0-4 branch

* patches/src680/jvmfwk-prefer-gij-on-x86_64.diff, patches/src680/apply:
(64bitPatches) prefer gij on x86_64 with the 64-bit build; It is
useful because Sun JRE 1.5 does not work with the 64-bit build at
all. It is caused by
http://bugs.sun.com/bugdatabase/view_bug.do?bug_id=6487640;
It is hard to debug. The wizards simply does not appear without any
message; The oobase functions ends with wery unclear message, for
example see http://forums.fedoraforum.org/showthread.php?t=135946,
n#219982

* patches/src680/jvmfwk-searched-paths.diff, patches/src680/apply:
# fixes searching for JREs; it allows to search subdirs, so all JREs
below /usr/lib/jvm are offered even when the first one is found in
/usr/lib, ...; it also pushes the right path, so one JRE is not
searched several times (fixes the the first very slow start);
it should be helpfull for everybody, n#203830, n#222708, i#64281

* patches/src680/jvmfwk-gij32.diff, patches/src680/apply: search also
gij32, ..., so the 32-bit package works on x86_64; It is NovellOnly
because I did not want to break others; it should be OK but ...
n#222708