Change cairo version used on Win32 to 1.
* configure.in: Change cairo version used on Win32 to 1.4.2.

* download.in: Download cairo sources from the official
cairographics.org site.

* patches/cairo/cairo-cairo-1-4-2-diff.diff: New file. Creates a
patch for cairo 1.4.2 that adds manually created config.h and
cairo-features.h files and makefiles for OOo's dmake and MSVC. The
MSVC changes to the asm code in the pixman library are now in
upstream cairo, so that is no longer needed.

* patches/cairo/cairo/cairo-makefile-mk.diff: Change tarfile
basename to cairo-1.4.2. Add one header to install:
cairo-deprecated.h

* patches/src680/apply: Use cairo-cairo-1-4-2-diff.diff instead of
cairo-cairo-1-2-4-diff.diff.