Add fix for n#300398 to cairo_canvashelper_text.
* patches/cairo/cairo-canvas-win32.diff: Add fix for n#300398 to
cairo_canvashelper_text.cxx: cairo does not keep the clip region
for the DC sensible for use outside cairo. Before displaying text
on the slide (which is not done through cairo) we need to force
cairo to restore the clip region with a dummy call to cairo to
fill an empty rectangle.

Thanks to Vladimir Vukicevic for the hint:
vlad_: tml_: the clip thing isn't really a bug
vlad_: cairo makes no promise that it will keep the clip up to date on the DC; it only does it when it needs to, and clip updating is lazy
vlad_: if you need to do stuff with the underlying DC, you can force the clip to be set... we settled on cairo_rectangle (cr, 0, 0, 0, 0); cairo_fill(cr);
vlad_: which has cairo doing the surface clip updating, and then it throws out the traps.. we pay a tessellation cost, but it's small