New folder structure for the Director.
* previous-idt-dirs: New folder structure for the Director.idt and
File.idt files from builds of previous versions, to be used when
building the installer on Windows so that it uses the same
mappings for File table keys and 8.3 names. (Search
Changelog-to-2.2 for the mention of PREVIOUS_IDT_DIR.) This is
necessary when doing builds to be used as targets for Windows
Installer patches, and also a good idea otherwise.

* previous-idt-dirs/Novell-OOo-2.1.0-12: New folder, for the
Novell "RTM" build of OOo 2.1.0.

* previous-idt-dirs/Novell-OOo-2.1.0-12/Director.idt
* previous-idt-dirs/Novell-OOo-2.1.0-12/File.idt: The actual files.