Don't return "com.
* patches/src680/novell-win32-odma.diff
(fpicker/source/odma/ODMAFilePicker.cxx): Don't return
"com.sun.star.ui.dialogs.FilePicker" as a supported service name
any longer. Causes problems in 2.4.0: the ODMA file picker is
invoked always, even if the option allowing it is turned off.

(scp2/source/ooo/file_ooo.scp): Temporarily surround the
additional stanzas for the template documents with #if 0, to avoid
error when constructing installer because the same file is in two
directories, or something like that. Correct fix is to use a
different NetDir for them.