Experimental patch to always enable save even when the document is not modified.
* patches/test/sfx2-always-allow-save-document.diff: experimental patch
to always enable save even when the document is not modified.  useful
when the view data changes (such as the current cursor location, zoom 
level etc.) but the user is unable to save it with the document because
such view data does not flag a document modified state.  Not enable yet
in the build.