Improve the fileversion program a bit: Make sure it always passes a full pathname to GetFileVersionInfo().
* patches/src680/novell-win32-msi-patchability.diff: Improve the
fileversion program a bit: Make sure it always passes a full
pathname to GetFileVersionInfo(). Should not be relevant for the
use of fileversion when building the OOo installer, as in that use
case it's already full pathnames that are passed on the command
line to fileversion.

But fileversion can be run interactively too, and this change
avoids strange surprises if one gives a name of a DLL in the
current directory and the same DLL happens to be found first in a
folder that is searched by GetFileVersionInfo() (or LoadLibrary())
before the current directory. Case in point: On XP gdiplus.dll is
found in a side-by-side folder even though a copy of it exists in
the current folder. Note that there is no gdiplus.dll in the
system32 folder, or anywhere in PATH.