The patch for boost/spirit/iterator/position_iterator.
* patches/src680/boost-patch.diff: The patch for
boost/spirit/iterator/position_iterator.hpp needs to be in
spirit-1.6.1.patch, not boost-1.30.2.patch. To avoid unnecessary
clarity and add some interestingness, the same header files are
unpacked both from the boost tarball and spirit tarball. It is the
spirit files from the spirit tarball that actually get delivered
and it is in there that we need to patch position_iterator.hpp.