Always draw a full border around the entire source selection, but use a different color than is used for the selected cell.
* patches/src680/sc-copy-source-border.diff: Always draw a full border
around the entire source selection, but use a different color than is used
for the selected cell.  This prevents the border from "disappearing"
immediately after a copy (which the prior logic tried to work around), and
I find it less ambiguous to use a different color for the copy source.