Latest code from koheidatapilot cws to fix a number of bugs and regressions (i#83250, i#88473).
* patches/src680/offapi-calc-datapilot-drilldown.diff:
* patches/src680/sc-calc-datapilot-drilldown.diff: latest code from
koheidatapilot cws to fix a number of bugs and regressions
(i#83250, i#88473).  This fixes regressions on drill-down sheet 
insertion when used with the 'Ignore empty rows' and/or 'Identify
categories' options, grouped fields, or numeric category names which 
must be evaluated numerically intead of texturally.  It also fixes
incorrect drill-down data when the field value is (empty) or 'Total'.

* scratch/sc-dptest/Makefile: adopted to DEV300 tree.