Act always as if --with-dejavu-fonts and --disable-build-mozilla had been used when on Windows.
* configure.in: Act always as if --with-dejavu-fonts and
--disable-build-mozilla had been used when on Windows. The DejaVu
fonts are de facto needed when building with ooo-build currently
on Windows. I haven't examined the root cause for this, but it
can't hurt to use them always anyway, right? The Mozilla 1.7.5
sources can't be compiled with MSVS2008 as of now. Either the
upstream mozilla-source-1.7.5.patch patch needs to be extended, or
we need to patch the Mozilla configury and Makefiles even more
ourselves.)

* download.in
* bin/unpack: Corresponding changes, and some
clarifications. Download the prebuilt Mozilla zipfiles. Comment
out the Mozilla source related bits for now. Once we can compile
Mozilla with MSVS2008 or some other solution to upstream issue
#66338 is in place will have to rework this.