Need a small patch also to config_office/configure.
* patches/src680/msvs9-temporary-fix.diff: Need a small patch also
to config_office/configure.in after all so that midl.exe and
mscoree.lib are found.

Still need to figure out what to do with the search for
instmsia.exe and instmsiw.exe. Those files are obsolete installers
for Windows Installer 2 and not even shipped any more with
MSVS2008 or Windows SDK 6.1. Probably it is pointless to including
them with the OOo installer.