Drop MSVS6, MSVC .
* patches/src680/msvs9-temporary-fix.diff
(config_office/oowintool): Drop MSVS6, MSVC .NET 2002 and MSVS
.NET 2002. Require the .NET Framework 2.0 ("2.0.50727") in
--csc-compilerdir (because cli_ure/source/basetypes/makefile.mk de
facto requires it, as it uses the -keyfile options which was not
in the 1.1 csc.exe). Correspondingly require .NET Framework SDK
2.0 in --dotnetsdk-dir. Copy also msvcm*.dll and
Microsoft.VC90.CRT.manifest in msvc_copy_dlls when using MSVS2008
or later.
(config_office/configure.in): Rely on oowintool to handle the
copying of the CRT DLLs and manifest. That's what oowintool is for
after all. Drop the attempt to do it here, including the horrible
repeated running of /bin/find. Still could do lots more cleaning
up of crap from here.