Added emf+-use-canvas-only-for-emf+.
* patches/src680/apply: added emf+-use-canvas-only-for-emf+.diff
with changes described below

* svtools/source/filter.vcl/wmf/winmtf.cxx: enable canvas
rendering for emf+/emf dual metafiles

* vcl/source/gdi/gdimtf.cxx: added bUseCanvas flag, which is used
during rendering, if set, the metafile is rendered with cppcanvas
metafile renderer instead of vcl renderer. it is set to false by
default, the reasoning is to improve performance for non emf+
metafiles as rendering thru canvas takes longer. it was cooked
together with Thorsten so that his new svg graphic import filter
works better