Don't unconditionally attempt to download the "system" tarball if building for Win32, only do it in the non-split case.
* download.in: Don't unconditionally attempt to download the
"system" tarball if building for Win32, only do it in the
non-split case. (But then, --with-split is de facto mandatory for
dev300-m* now.) Don't unconditionally attempt to download the
"binfilter" tarball if binfilter is enabled, it shouldn't be
needed in the --with-split case, says Rene. Download the
libs_extern_sys tarball also for NovellWin32 and GoOoWin32. Fix
spelling of "libs_extern_sys" in the non-split case of mapping it
to "system".