Add DejaVuFonts to NovellLikeWin32.
* patches/dev300/apply: Add DejaVuFonts to NovellLikeWin32. I
don't really like that this patch set then also makes the
Liberation fonts mandatory, but still one has to explicitly
specify --with-liberation-fonts. That should be implicit then,
like --with-dejavu-fonts. 

Do we want to make DejaVu and/or Liberation fonts mandatory on all
Win32 builds, or just for NovellLikeWin32 and
OxygenOfficeWin32Common? (Not that I know if there are any others
actively building ooo-build for Win32 than Novell and
OxygenOffice.)

This is a mess. The knowledge about what functionality is enabled
in what "distros" and what needs to be downloaded and/or copied
into the build tree is split between configure.in, bin/unpack and
patches. And still some things have to be explicitly specified on
the configure command line. Sigh.