Allow users to  change document's current passowrd for encryption, so that they won't have to save as and overwrite the current document to change it.
* patches/dev300/document-password-change-sfx2.diff: allow users to 
change document's current passowrd for encryption, so that they won't
have to save as and overwrite the current document to change it.  When
the current password is changed, a simple save will save it with the 
new password.  To change password, go to File -> Properties and on the
General tab, there is a button named "Change Password...".

* patches/dev300/min-password-length-sfx2.diff: reduce the mininum 
password length required for odf encryption from 5 to 1.  A single
letter password works just fine, and there is no justification (other
than that "short passwords are insecure" mantra) for requiring a length
of 5 (i#85453).

* patches/dev300/apply: apply these new patches.