Drop most of our stuff from this file.
* patches/src680/novell-win32-msi-patchability.diff: Drop most of
our stuff from this file. So much has changed in the details of
how upstream produces the installer, and the MSI patching using
our builds didn't work out very well the way I tried do handle it
in 2.4.1 anyway, so try a fresh start based on how upstream does
it.