Merge from ooo-build-3-0:
Merge from ooo-build-3-0:

* patches/dev300/novell-win32-odma.diff
(fpicker/source/odma/ODMAFolderPicker.cxx): Don't claim we support
the service com.sun.star.ui.dialogs.FolderPicker. Unlike for the
corresponding file picker, it isn't necessary. It's enough to just
say that we support the com.sun.star.ui.dialogs.ODMAFolderPicker
service.

This fixes the unintended phenomenon that even if one toggles on
the "Use OpenOffice.org dialogs" on the Tools:Options:General
page, it is still the ODMAFolderPicker (which forwards to the
Win32FolderPicker in the usual case of no ODMA server available)
that gets used. This breaks testtool scripts. When the system
folder dialog is displayed, it is modal and sticks, and testtool
loses track of what is going on and everything fails after that.