Merge from ooo-build-3-0:
Merge from ooo-build-3-0:

* patches/dev300/win32-tooltips.diff: Just drop the
ALL_LANG(Tooltip, STR_FI_TOOLTIP_FOO) entries from the scp files,
so that no tooltips are created in the first place, instead of
trying to make the tooltips less lame and less misleading. In
particular, drop the misleading "Supported by Sun Microsystems"
from the icon on the desktop. They are localised anyway, and I
don't want to attempt to edit them in umpteen languages in the
localize.sdf file.

* patches/dev300/apply: Add win32-tooltips.diff to Fixes.