Take out tests for unicows.
* patches/dev300/novell-win32-drop-win9x.diff: Take out tests for
unicows.lib and instmsia.exe from configure.

* patches/dev300/cygwin-make-ver.diff: Revert breakage. Now the
test again checks if we are using make 3.81 and tells us to use
3.80 in that case. In the previous version it checked if we use
make 3.80, and in that case told us to use 3.80...

As such, if the Make 3.81 problem really is only related to
building Mozilla, this patch would not be needed for now as we
don't build the Mozilla bits on Windows anyway because the Mozilla
version in question can't be built with MSVS9.