Now it seems to work again.
* patches/dev300/webdav-locking-from-ooo-build-2-4-1.diff: Now it
seems to work again. Basically what I did in the end, once I had
an understanding of what is going on, was to move the periodical
relocking and the unlocking of the WebDAV resource from
webdav_ucp::NeonInputStream to webdav_ucp::Content. This was
necessary because the NeonInputStream is live only while the
document is being loaded. Presumably this was different then in
2.4.1 and 3.0.x? Was that because of upstream code, or did the
other patches for WebDAV in those versions make it so? Anyway, I
didn't understand how to change that back now in 3.1, or whether
that would even be a good idea.

I don't know exactly what effect the SupportsActiveStreaming stuff
exactly is supposed to have, and whether it in fact is necessary
or not now that it's webdav_ucp::Content that manages both locking
and unlocking.

A sidenote: In my last testing against the Teaming innerweb
server, WebDAV locks with timeouts did not expire at all... But
presumably that is then just a bug in the teaming server or its
configuration. Or maybe an intentional tweak by the site's
admins to avoid some class of problems? After all, the Teaming UI
does allow unlocking a document explicitly. So presumably if some
document would remain locked because the software used to edit
it crashed or otherwise didn't unlock it, and the lock doesn't
time out by itself, that is not a serious problem.