There were still problems in the lock renewal code.
* patches/dev300/webdav-locking-from-ooo-build-2-4-1.diff: There
were still problems in the lock renewal code. I don't know if it
is a good idea at all to call ucbhelper::cancelCommandExecution()
in the ticker thread? Seemed to cause a crash for me. Also, the
code assumed that the error message retrieved by ne_get_error()
always starts with a numeric code, which definitely is not
true. Case in point: the "No activelock " error.

This error seems to happen regularily at least when trying
ne_lock_refresh() against the Teaming server, so we check
explicitly for that message. In such cases we then just forget the
existing lock and create a new one, instead of calling
cancelCommandExecution().