Stop adjusting cases based on values in the  same column or values in the sort list.
* patches/dev300/calc-autoinput-case-insensitive-matching.diff:
* patches/dev300/apply: Stop adjusting cases based on values in the 
same column or values in the sort list.  This prevents, for example, 
capitalized Dutch month names from being entered correctly.  As a 
consequence of this change, however, entering for example 'jan' in the
en-US locale will no longer get auto-converted to 'Jan', but hopefully 
that won't be a big deal. (n#472395, lp#271774)