Allow ctrl-[ and ctrl-] to jump to precedents and dependents of a cell.
Similar to Excel, allow the users to jump to precedents and dependents of an
arbitrary formula cell, via ctrl-[ and ctrl-], respectively. (n#464359)

Since this patch set introduces new key input types (KEY_BRACKETLEFT and
KEY_BRACKETRIGHT), we need to test it in different desktop environments.  I
have already tested it in GNOME, but not in KDE, pure X, Windows, Mac OS X
and possibly others.

* patches/dev300/apply:
* patches/dev300/calc-jump-on-formula-ref-offapi.diff:
* patches/dev300/calc-jump-on-formula-ref-officecfg.diff:
* patches/dev300/calc-jump-on-formula-ref-sc.diff:
* patches/dev300/calc-jump-on-formula-ref-sfx2.diff:
* patches/dev300/calc-jump-on-formula-ref-vcl.diff:

* patches/test/calc-jump-on-formula-ref.diff: backup patch removed.