Fixed infinite loop on exporting ODS document.
When a document, created with 3.0 with row styles spanning over to the
maximum row size, got saved in 3.1, it would hang when Calc tried to
save its automatic styles.  The hang was due to the row limit increase,
and the fix is to check for the last edited row, and don't save beyond
that last edited row to prevent an infinite loop. (n#497610)

* patches/dev300/apply:
* patches/dev300/calc-odf-export-autostyle-maxrow.diff: