OOXML: Create separate libxlsx and libpptx (similarly to libdocx).
For the purpose of the 3.1 release, instead of sharing the binary and ooxml
export, we introduce a separate sc/source/filter/xlsx and
sd/source/filter/pptx that creates libxlsx & libpptx.

After the release, the changes will be reverted so that the binary and ooxml
filters share the code again.

* patches/dev300/apply:
* patches/dev300/pptx-export-disable-animations.diff:
* patches/dev300/pptx-export-empty-animations-node-and-groupshape-fix.diff:
* patches/dev300/pptx-filter-as-a-separate-lib.diff:
* patches/dev300/pptx-snapshot-at-ooxml03-creation.diff:
* patches/dev300/pptx-the-rest-from-git.diff:
* patches/dev300/xlsx-arabic-export-crash.diff:
* patches/dev300/xlsx-filter-as-a-separate-lib.diff:
* patches/dev300/xlsx-shared-import-and-export.diff:
* patches/dev300/xlsx-snapshot.diff: