Changes to the git initialization script
The goal of those changes is to handle applied_patches/* in a local
git repo with the sources for easier use of the makefile when switching
to a branch with a different set of applied patches.

In order to be able to produce nice patches, these applied_patches/*
are falling into a black hole when producing diffs.

* bin/create-gitignores.sh:
    - Creates a build/bin/nothing script
    - Creates a .gitattributes
    - Appends a diff driver to .git/config
    - Don't set /applied_patches in .gitignores