Fix the fact that rotated groups from PPT are sometimes misplaced
The bug n#485637 has a doc that, when imported, has rotated groups
(the arrows in this case) distorted & misplaced. That's due to the
fact that svx/source/msfilter/msdffimp.cxx tweaks the wrong bound
rect when trying to adapt it to rotation (angles from {[45-135],
[225-315]} degrees apparently need swapping x & y) - instead of the
outer group bounds, the intra-group coordinate system was swapped,
but that was already in a correct state.

* patches/dev300/apply: added the patch
* patches/dev300/svx-pptin-global-child-bounds-fix.diff: fixed as
  outlined above