Minor tweaking of pagination.
The increment of nY at the end of each loop is probably not needed.
In the original code, the nY value was taken from the iterator in each
iteration, so incrementing it at the end had no effect in the next
iteration.  But the new code uses a for-loop.  So, it's probably not
a good idea to manually increment it.

* patches/dev300/calc-perf-page-and-manual-breaks.diff: