Make .pps/.ppsx files autoplay & close doc afterwards
This is a fix for n#485645 and makes .pps and .ppsx (the ooxml
variant of the ppt auto-show files) start the slideshow after
loading, and closing the document automatically after the slideshow
has ended. When the .pps/.ppsx doc was the only document open, OOo
will shut down as usual.

* patches/dev300/apply: added pptx-autoplay-fix.diff, removed
  sd-pps-autoplay.diff (which was a rather lame and hacky attempt at
  enabling autoplay for .pps files)
* patches/dev300/pptx-autoplay-fix.diff: the proper fix
* patches/dev300/sd-pps-autoplay.diff: removed