Fixed Impress' way of numbering paragraphs
With 3.0, Impress numbering was modified because of the odf 12
changes in list level handling; we differ in the way we
consider paragraphs to belong to the 'same' numbering group,
compared to ppt. now, as long as the para has the same outline
level (and no 'restart numbering' flag), it is considered to
belong to the same numbering group, i.e. numbering will mono-
tonically increase. Compares favorable to PPT. Additionally,
fixed a repaint bug that prevented following paragraph's numbers
to change, when certain modifications were made on the current one.

* patches/dev300/sd-ruler-fix.diff: updated as outlined above