Set the lower threshold for selection engine's auto repeat interval.
On Windows, setting the auto repeat interval too low (i.e. ~5) caused
the scrolling during range selection to get stuck, while on Linux we
don't have this problem.  I don't really know if this is a system
limitation or something we do in vcl, but setting the lower threshold
to 25 seems to fix this problem on Windows. (n#514395)

* patches/dev300/scroll-accel-vcl.diff: